requirements.txt
asgiref==3.5.2
Django==4.0.5
mysqlclient==2.1.0
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
django-rest-knox==4.2.0
